fix: split validation inside validations folder
