+++ server/validations/auth.js
@@ -1,4 +1,4 @@
-const { check, body, validationResult } = require("express-validator");
+const { check, validationResult } = require("express-validator");
 
 exports.validateRegister = [
   check("username", "Please enter a username").not().isEmpty(),
@@ -22,22 +22,6 @@ exports.validateRegister = [
 exports.validateLogin = [
   check("email", "Please enter a valid email address").isEmail(),
   check("password", "Password is required").not().isEmpty(),
-  (req, res, next) => {
-    const errors = validationResult(req);
-
-    if (!errors.isEmpty())
-      return res.status(400).json({ errors: errors.array() });
-    next();
-  }
-];
-
-
-exports.validateCardEdit = [
-  check("tags", "Tags must be send in an array.").isArray(),
-  check("tags.*", "Tags must be a string.").isString(),
-  check("attachments", "Attachments must be send in an array.").isArray(),
-  check("attachments.*", "Attachments must be a string.").isString(),
-  check("colorCode").isString(),
   (req, res, next) => {
     const errors = validationResult(req);
